requirement.txt
absl-py==2.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
Brotli==1.0.9
certifi==2025.8.3
charset-normalizer==3.3.2
click==8.3.1
comm==0.2.3
contourpy==1.3.1
cycler==0.11.0
debugpy==1.8.17
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.128.0
filelock==3.17.0
fonttools==4.55.3
gmpy2==2.2.1
grpcio==1.71.0
h11==0.16.0
httptools==0.7.1
idna==3.7
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.8
lazy_loader==0.4
Markdown==3.4.1
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.2.1
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mpmath==1.3.0
MultiScaleDeformableAttention==1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.3.2
parso==0.8.5
patsy==1.0.2
pexpect==4.9.0
pillow==11.1.0
pip==25.3
platformdirs==4.5.1
polars==1.37.1
polars-runtime-32==1.37.1
prompt_toolkit==3.0.52
protobuf==5.29.3
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.0
PyQt6==6.7.1
PyQt6_sip==13.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.2
setuptools==80.10.1
sip==6.10.0
six==1.17.0
stack-data==0.6.3
starlette==0.50.0
statsmodels==0.14.5
sympy==1.14.0
tensorboard==2.19.0
tensorboard_data_server==0.7.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==2.0.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
ultralytics==8.4.7
ultralytics-thop==2.0.18
unicodedata2==15.1.0
urllib3==2.3.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==16.0
Werkzeug==3.1.3
wheel==0.46.3
